Impl IntoArma for missing integer type
Added these since they do have FromArma implemented but not vice versa.
Most of these are likely to run into floating point precision issues,
but the way I see it this is more a problem at large since this also
already an issue with `i32` and `u32` all be it not as likely.